+++ tests/testthat/test-model_frame.R
@@ -1,4 +1,4 @@
-if (require("testthat") && require("sjstats") && require("splines")) {
+if (require("testthat") && require("sjstats") && require("splines") && require("glmmTMB")) {
   context("sjstats, model_frame")
 
   data(efc)
@@ -22,6 +22,21 @@ if (require("testthat") && require("sjstats") && require("splines")) {
     expect_equal(as.vector(mf2$c160age), as.vector(mf5$c160age))
     expect_equal(as.vector(mf3$c160age), as.vector(mf5$c160age))
     expect_equal(as.vector(mf4$c160age), as.vector(mf5$c160age))
+  })
+
+  data("Salamanders")
+  m <- glmmTMB(
+    count ~ spp + cover + mined + poly(DOP, 3) + (1 | site),
+    ziformula = ~spp + mined,
+    dispformula = ~DOY,
+    data = Salamanders,
+    family = nbinom2
+  )
+
+  test_that("model_frame", {
+    mf <- model_frame(m)
+    expect_equal(ncol(mf), 7)
+    expect_equal(colnames(mf), c("count", "spp", "cover", "mined", "site", "DOY", "DOP"))
   })
 }